Add small code chagnes

+++ src/gamesound.cpp
@@ -1,7 +1,7 @@
-extern void set_volume(double, int);
+extern void set_volume(float, int);
 
 void parse_sound_from_sounds_tbl()
 {
     char buffer[264];
-    set_volume((double)0x80228f9c,0);
+    set_volume((float)0x80228f9c,0);
 }
